// src/type.d.ts
export declare const typeObject: {
    function: string;
    date: string;
    regExp: string;
    error: string;
    map: string;
    set: string;
    weakMap: string;
    weakSet: string;
    promise: string;
    arrayBuffer: string;
    int8Array: string;
    uint8Array: string;
    uint8ClampedArray: string;
    int16Array: string;
    uint16Array: string;
    int32Array: string;
    uint32Array: string;
    float32Array: string;
    float64Array: string;
    bigInt64Array: string;
    bigUint64Array: string;
    dataView: string;
    blob: string;
};
export declare const TypeValue: {
    unknown: "unknown";
    function: "function";
    date: "date";
    regExp: "regExp";
    error: "error";
    map: "map";
    set: "set";
    weakMap: "weakMap";
    weakSet: "weakSet";
    promise: "promise";
    arrayBuffer: "arrayBuffer";
    int8Array: "int8Array";
    uint8Array: "uint8Array";
    uint8ClampedArray: "uint8ClampedArray";
    int16Array: "int16Array";
    uint16Array: "uint16Array";
    int32Array: "int32Array";
    uint32Array: "uint32Array";
    float32Array: "float32Array";
    float64Array: "float64Array";
    bigInt64Array: "bigInt64Array";
    bigUint64Array: "bigUint64Array";
    dataView: "dataView";
    blob: "blob";
    null: "null";
    undefined: "undefined";
    NaN: "NaN";
    infinity: "infinity";
    integer: "integer";
    float: "float";
    string: "string";
    class: "class";
    generator: "generator";
};